refactor(moonrun): separate runtime implementation layers
